Did not notice the existing test class. Moving my tests there.
